requirements.txt
python-gitlab
python-dotenv
google-generativeai
openai
pyyaml